move jwt logic out of base hook
